Response data
Response data from mongoDB to Request

+++ Back/Model/ConvidCasesStatistic.go
@@ -10,12 +10,12 @@ type ConvidCasesStatistic struct {
 
 type StatisticsEachDay struct {
 	Date            string `json:"Date" gorm:"Date"`
-	NewConfirmed    int16  `json:"NewConfirmed" gorm:"NewConfirmed"`
-	NewRecovered    int16  `json:"NewRecovered" gorm:"NewRecovered"`
-	NewHospitalized int16  `json:"NewHospitalized" gorm:"NewHospitalized"`
-	NewDeaths       int16  `json:"NewDeaths" gorm:"NewDeaths"`
-	Confirmed       int16  `json:"Confirmed" gorm:"Confirmed"`
-	Recovered       int16  `json:"Recovered" gorm:"Recovered"`
-	Hospitalized    int16  `json:"Hospitalized" gorm:"Hospitalized"`
-	Deaths          int16  `json:"Deaths" gorm:"Deaths"`
+	NewConfirmed    int32  `json:"NewConfirmed" gorm:"NewConfirmed"`
+	NewRecovered    int32  `json:"NewRecovered" gorm:"NewRecovered"`
+	NewHospitalized int32  `json:"NewHospitalized" gorm:"NewHospitalized"`
+	NewDeaths       int32  `json:"NewDeaths" gorm:"NewDeaths"`
+	Confirmed       int32  `json:"Confirmed" gorm:"Confirmed"`
+	Recovered       int32  `json:"Recovered" gorm:"Recovered"`
+	Hospitalized    int32  `json:"Hospitalized" gorm:"Hospitalized"`
+	Deaths          int32  `json:"Deaths" gorm:"Deaths"`
 }
